requirements.txt
flask
flask-cors
python-dotenv
gunicorn==21.2.0
faster-whisper==1.0.2